Allow for a 4 char suffix on the end of hostnames (#156)
* Allow for a 4 char suffix on the end of hostnames

As we migrate cloud deployments to use a load balancer backed by managed
instance groups, we can no longer count on machine names to be
predetermined. MIG instances have a "-<4 char>" suffix, and are managed
by the instance group, not M-Lab. In order to support this use case, the
changes in this PR allow for a suffix to be part of the hostname without
failing Parse(). Callers can use the suffix, if found, or not.

* Adds example of MIG instance name with rand suffix