Adds funcs to return strings w/ prefixes and suffixes
Previously, the only string function returned a typical mlab hostname.
This commit adds additional string functions to return various
combinations of the hostname with or without prefixes and suffixes.

Additionally, this commit modifies the v2 regexp to capture and return
an prefix. Previously the prefix match was a non-capturing group.