Fixes String functions + adds more unit tests
I realized that some of the String* functions would return bad hostnames
in the event that certain fields were not present in the input hostname.
This commit should fix that.

Also, this commit adds unit tests for the new String* functions to be
sure they are doing what we expect.